+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## [v0.0.33]
+- null checking choices[0]
+
+
+## [v0.0.32]
+- handle null chunk.choices idk why
+
+## [v0.0.31]
+- fix tool description truncation for openai models #45
+
 ## [v0.0.30]
 - support deepseek api reasoning model's reasoning content